+++ R/col_types.R
@@ -4,10 +4,12 @@
 #' as the default. `cols_only()` includes only the columns you explicitly
 #' specify, skipping the rest.
 #'
-#' @param ... Either column objects created by `col_*()`, or their
-#'   abbreviated character names. If you're only overriding a few columns,
-#'   it's best to refer to columns by name. If not named, the column types
-#'   must match the column names exactly.
+#' @family parsers
+#' @param ... Either column objects created by `col_*()`, or their abbreviated
+#'   character names (as described in the \code{col_types} argument of
+#'   \code{\link{read_delim}}). If you're only overriding a few columns, it's
+#'   best to refer to columns by name. If not named, the column types must match
+#'   the column names exactly.
 #' @param .default Any named columns not explicitly overridden in `...`
 #'   will be read with this column type.
 #' @export
@@ -15,7 +17,7 @@
 #' cols(a = col_integer())
 #' cols_only(a = col_integer())
 #'
-#' # You can also use the standard abreviations
+#' # You can also use the standard abbreviations
 #' cols(a = "i")
 #' cols(a = "i", b = "d", c = "_")
 #'
@@ -219,6 +221,7 @@ str.col_spec <- function(object, ..., indent.str = "") {
 #' `spec()` extracts the full column specification from a tibble
 #' created by readr.
 #'
+#' @family parsers
 #' @param x The data frame object to extract from
 #' @return A col_spec object.
 #' @export