+++ NEWS.md
@@ -33,6 +33,7 @@ you would pass `skip = 3`, now you only need to pass `skip = 2`.
 * `write_file()` gains a `sep` argument, to specify the line separator (#665).
 * Allow files to be read via FTP over SSH by recognising `sftp` as a URL protocol (#707, @jdeboer).
 * `parse_date*() accepts `%a` for local day of week (#763, @tigertoes).
+* Added function `read_lines_raw_chunked()` (#710, @gergness)
 
 ## Bug Fixes
 
@@ -50,6 +51,7 @@ you would pass `skip = 3`, now you only need to pass `skip = 2`.
   integer values in the first 1000 (#645, #652).
 * `read_*()` now converts string `file`s to UTF-8 before parsing, which is convenient for non-UTF-8 platforms
   in most cases (#730, @yutannihilation).
+* `write_csv()` writes integers up to 10^15 without scientific notation (#765, @zeehio)
 * `read_*()` no longer throws a "length of NULL cannot be changed" warning when
   trying to resize a skipped column (#750, #833).
 * `read_*()` now handles non-ASCII paths properly with R >=3.5.0 on Windows (#838, @yutannihilation).